requirements.txt
flask==2.3.2
pillow==10.0.0
requests==2.31.0